Style: SignUp 컴포넌트에 email 과 gender radio box 부분 컴포넌트로 분리
